SIM-7918 Text Decluttering
Release Notes: simVis::RangeTool now uses the midpoint of the slant range for all text.


Notes:  This is the first pass that moves the text to the midpoint of the slant range.  I also kept getting an assert on clampEcefPointToGeodeticSurface when a vector of all zeros is passed in.  I would get the assert when turning on the overhead mode.


Test Environment:  Windows, Visual Studio 2013, Debug, 64-bit


Test: Manual Testing with simple server.


Review: 12248