Improved the simCore::Mgrs::convertUpsToGeodetic function
SDK REL: Improved the performance and robustness of simCore::Mgrs::convertUpsToGeodetic function.

convertUpsToGeodetic function was replaced with an adaptation of the GeographicLib PolarStereographic::Reverse() function.

The function is no longer iterative and is much faster.

The hemisphere enumeration was also replaced with a boolean value to denote North (true) vs. South (false) pole.  Several of the function signature have also been updated.