Updated getDeltaT function to use new coefficients
SDK REL: The getDeltaT function has been updated to use newer time coefficients for the prediction of the delta T value.

Comparison of the predicted delta T value to tabular data from USNO indicated the value was off by over 2 seconds.  The USNO published new time coefficients in 2012.  When those values are used, the predicted difference is now off by less than .2 seconds versus the actual delta T value.

Note that the predicted value is used since the actual value is based on the determination of whether a leap second is added to the astronomical clock.  In 2018 is was determined that a leap second was not needed.

13056