Performance Update to LOB with Tables
Release Notes: N/A


Notes:  Loading a file with 50,000 LOB and 50,000 Tables went from 120 seconds to 10 seconds.   In both paths an extra std::map::find call is added if the table simData::INTERNAL_LOB_DRAWSTYLE_TABLE exists.  In the routine setLineDrawStyle_ an extra call is added if the table does not exist, but the routine is only called when a new LOB line is created. This approach spreads out the cost of LOBs and tables to the active LOBs, which should be considerably less than 50,000, and should not be noticeable.


Test Environment:  Windows, Visual Studio 2013, Debug, 64-bit


Test:  ASI file with 50,000 LOB and 50,000 Tables; Simple Server Plug-in


Review: 13993