SIM-8607 Updated Compile Options
Release Notes: N/A


Notes:  SIMDIS 10 could not compile on Windows without the –bigobj when the custom rendering was selected.  I placed the option outside the ENABLE_CUSTOM_RENDERING, since the error indicates we are on the edge with respect to size.


Test Environment:  Windows, Visual Studio 2013, Debug, 64-bit


Test:  Compile SIMDIS 10


Review: 14077