SIM-8700: Update Entity State Filter (1 of 2 SDK)
Release Notes: N/A


Notes:  The entity state filter now checks the life span of the custom rendering entity.  A method was added to allow for the editing of custom rendering commands to that the life span can be modified when the persistence value changes.


Test Environment:  Windows, Visual Studio 2013, Debug, 64-bit


Test:  Manual testing with SIMDIS 10.


Review: 14228