SIM-8679 Added Custom Rendering to simVis::RangeTool
Release Notes: N/A


Notes:  The screen shot has which calculations are enabled when a custom rendering is the first entity or the second entity.  A custom rendering could have a velocity, but the velocity is currently not available to the simVis::RangeTool.  A special path was done for platforms and we would need to do the same thing if velocity is required for simVis::RangeTool.  I also fixed a bug in isBeamToEntity_.


Test Environment:  Windows, Visual Studio 2013, Debug, 64-bit


Test:  Manual testing.


Review: 14359