SIM-8696 Local Grid for Custom Rendering
Release Notes: N/A


Notes:  I also fixed a bug in Local Grid.    The “Use Platform Speed” does not make sense for Custom Rendering, but I was not planning to address it.   I also fixed a bug where the name for the label was always using the lastPrefs_ while the rest of the fields came from the current preference.  After this review I was going to do a cleanup pass on all the entity types and implement a similar fix.


Test Environment:  Windows, Visual Studio 2013, Debug, 64-bit


Test:  Tested various grid options and they all worked


Review: 14391