SIM-8702 Update SDK Custom Rendering Example
Release Notes: N/A.


Notes:  Replaced the triangle with a constantly changing unit circle.  I updated a method signature.   The Entity label was failing on the first update and if the text did not change after the first update the error was very noticeable.  Applying the text at the end of the routine fixes the problem.  Update the documentation.


Test Environment:  Windows, VC12, Debug, 64-bit


Test:  See screen shot.


Review: 14890