SIM-8968 Added warning when LOS cannot be generated
Release Notes: N/A.


Notes:  I added some code to cut down on the spam.   If we want to add platform name to the error message it will be necessary to add a dependency on the platform node or refactor several RadialLOSNode methods to return an error.   On this pass I went with the cheap solution.


Test Environment:  Windows, VC12, Debug, 64-bit


Test:  Manual testing with SIMDIS-10.


Review: 14945