SPR-1724 Refactor RangeToolState Class (1 of 2 SDK)
Release Notes: N/A.


Notes: This is the third of many steps.  The new classes are in the original files to make the review easier, but I plan to break the new classes out to new files once this review passes.  There are many changes, but most were global search and replace.


Test Environment:  Windows, Visual Studio 2013 Debug, 64-bit


Test:  SDK Range Tool Example still works.


Review: 15474