SIM-9979 Case Sensitivity Option for the Category Manager (1 of 2 SDK)
Release Notes: A flag was added to control if the simData::CategoryNameManager is case sensitive.


Notes:   By default the manager is case sensitive which matches the current behavior.  If the manager is switched to case insensitive then the first appearance of a string value sets the case.  If string values of LINK16 and Link16 are given to the Category Manager then the value kept by the Category Manager is LINK16, since it was the first string value received by the Category Manager.


Test Environment:  Windows 10, Visual Studio 2013, Debug, 64-bit


Test:  Updated unit test and manual testing with SIMDIS.


Review: 17977