SIM-10292 New Binding for simQt::CenterEntity (1 of 2 SDK)
Release Notes: simQt::BindCenterEntityToEntityTreeComposite was added to provide full binding between simQt::CenterEntity and simQt::EntityTreeComposite.


Notes:  The enable state for the Center feature was not always correct.  The root cause of the problem was the enable state was calculated when the selection changed.  However, the state is time dependent and needs to be calculated just before the menu is displayed.  I need to go back and make the same change to the PluginUtils::QtCenterEntity I made yesterday. The next pass will look at updating simQt::EntityLineEdit.


Test Environment:  Windows 10, Visual Studio 2013, Debug, 64-bit


Test:  Manual testing with the Data Browser, Pref Tool and others.


Review: 18604