SIM-10345 Centering on Inactive Custom Rendering Will Change Time to an Active State
Release: N/A


Notes:  This is the last pass on this task.  There is one aspect of this code I do not like.  I want to minimize the time change, so the code backs up to the end of the previous time range. The end time is not valid so the code backs up an additional microsecond to get a valid time. I am open to other ideas.


Test Environment:  Windows 10, Visual Studio 2013, Debug, 64-bit


Test:  Manual testing with a small ASI file.


Review: 19055