SIM-10591 More simQt::TimeWidget Fixes
Release: N/A.

Notes:  A display precision lower than the data precision expands the range so that the widget does not report errors.  For example if the start time is 0.55 seconds but the precision set to 1, the widget will show 0.5.  Before this change the widget would return 0.5 for the time.  The widget now limits against the actual range instead of the display range.  Testing this change showed that the display range was not updated when the precision was changed.

Test Environment:  Windows 10, Visual Studio 2015, Debug, 64-bit

Test:  Manual testing with the simQt::TimeWidget test application.

Review: 19587