SIM-10602 Added Visualization Clock (1 of 3 SDK)
Release: N/A.

Notes: There needs to be a visualization clock and a data clock.  I combined them behind one simCore::Clock interface.  A Boolean controls which clock to use.  Most code will not know the difference.  There is some code that have early kicks outs if in Live Mode, the kick outs need to be removed and this class needs to decide if the request should be ignored.

Test Environment: Windows 10, Visual Studio 2015, Debug, 64-bit

Test: Updated and ran SIMDIS-X

Review: 19603