SIM-10602 Fixed Missing Platforms
Release: N/A.

Notes: The Data Store was hiding platforms because it thought the clock was in File Mode.  In Replay the clock now always returns true for Live Mode.  I did not see this issue with Simple Server Testing because of the high data rate.

Test Environment: Windows 10, Visual Studio 2015, Debug, 64-bit

Test: SIMDIS-X with ReadUMF, platforms no longer disappear when the replay time is close to the data time.

Review: 19614