SIM-11163 Removed Invalid Assert in simCore::calculateGeodesicDRCR
Release: N/A.

Notes:  I wrote my new approach and got basically the same answers.  The approach by simCore::calculateGeodesicDRCR is correct.  The angle between the Down Range line and the Cross Range line is 90 degrees.  What is wrong is the assert that the angles of the triangle equals 180 degrees.  The sum of the angles of a triangle on a sphere will be larger than 180 degrees.  See: https://en.wikipedia.org/wiki/Spherical_geometry for details.

Test Environment: Windows 10, Visual Studio 2017, Debug

Test:  All unit test still pass.

Review: 20873