SDK: SIM-13499 Hopefully A More User Friendly Entity List (1/2)
Release Notes: N/A

Notes: An attempt to make the Entity list less annoying when entities are added or removed from the list.   If there is a selected item and the item is in view then item will stay in view after the add or delete.  I also removed an undefined declaration.  Consider building all of VSI, to makes sure dynamic and static users of the simQt are updated.

Test Environment: Windows, Visual Studio 2019, Debug

Test: Ran with Simple Server and Scenario Editor.  I manually added/deleted/renamed entities and the results looked good.

Review: 26968