SIM-13922 Fixed Accepts Projector When Time Resets Accepts Projector To None
Release Notes: N/A.

Notes: The code was not correctly resetting when time went back to the start or when a new scenario was loaded.   I got rid of the code that scans a generic message looking for repeated fields and added code just update the one repeated field.

Test Environment: Windows 10, Visual Studio 2017, Debug

Test: Updated and ran AcceptsProjectors.sts

Review: 27187