Expand the Data Displayed in Scenario Properties (1 of 2)
**Release Notes:** 
simData::DataStore::idCount() is a new function that returns the number of unique entities of a given entity type.

**JIRA Issue:** 
SIM-15435

**Description:** 
SDK changes needed to add a method to return entity counts from the data store.

**Notes:**
Platform counts shown in the Scenario Data dialog utilized the method that returns a list of IDs, then used the size of the list to show the value.  Instead of returning lists and requesting a size, a new method that just returns the count is now used.

**Testing Performed:**
SIM-15435 changes in SIMDIS now use this method to show entity specific counts as well as a total
Compiled vsi.sln, no errors or warnings