SDK: simQt DockWidget handles restoring geometry better when no settings exists, SIM-18949
**Release Notes** simQt::DockWidget handles saving and restoring geometry better if there is no Settings object

**JIRA Issue** SIM-18949
**Summary** SDK: simQt DockWidget handles restoring geometry better when no settings exists
**Description** Plug-ins that create a simQt DockWidget do not have access to the main simQt Settings, so they cannot save/load the geometry from the settings. This only manifests in specific circumstances: if the plug-in dialog is closed and deleted without ever being docked, and before the QMainWindow serializes state.
Once a QDockWidget becomes known to the QMainWindow, it is automatically tracked as part of that QMainWindow's state, even if it is floating. By docking the widget momentarily, this allows the QMainWindow to keep track of its geometry.

Note that S060 step 4 would fail with certain plug-in dialogs that delete on close prior to this change. Since not all plug-ins delete on close, it looks like our regression testing always happened to use plug-in dialogs that don't

**Testing** ran S060, S061, and S062 regression tests, tested with Chat Display plug-in (which deletes on close) and P5 dialog (which does not), as well as internal dialogs (prefs tool, data browser).